Create colorpicker
This commit adds an elm file, Controller.elm, which allows you to select
a color by moving your mouse around. X axis is hue, Y is saturation. If
you click, Y will change lightness. Plan is to send this value to the
Elixir app, which will then relay it to the phones.