Rewrite as a phoenix app (#1)
* Initial Phoenix commit

* Sync colors

This gets the color syncing actually working through Phoenix channels.

The Elm code is pretty ugly around this... the clients need to
subscribe to a specific channel before they start receiving messages.
It's a nice abstraction! But I don't need it in this app.

* WIP: heroku

* Only display connected users if we have that info

Doesn't really help to show "0 connected users" right now, because the
Phoenix presence stuff is not currently enabled.

* Use system's port

* Use wss for websocket connection

* Parameterize socket

I hard-coded the socket just to get this thing up and running on heroku,
but it wasn't too tricky to pass it into the elm program with a flag.
Hardest part was generating the websocket url. Crazy that there's not
a good standardized `socket_url` helper in phoenix.

* Remove some brunch cruft

* Add script/server and update readme

Convention to install dependencies and start the dev server.